# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lwang Black API",
    "description": "E-commerce backend API for Lwang Black specialty coffee. Supports multi-currency payments (Stripe, PayPal, Khalti, eSewa, Nabil Bank), global logistics (Shippo, DHL, FedEx), admin dashboard, and notification services.",
    "version": "1.0.0",
    "contact": {
      "email": "brewed@lwangblack.co",
      "url": "https://www.lwangblack.co"
    }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development" },
    { "url": "https://api.lwangblack.co", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "qty": { "type": "integer" },
          "price": { "type": "number" }
        }
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "fname": { "type": "string" },
          "lname": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "country": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List orders (filtered by role/country)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "country", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "Paginated order list" } }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create a new order",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "country": { "type": "string" },
                  "currency": { "type": "string" },
                  "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItem" } },
                  "subtotal": { "type": "number" },
                  "shipping": { "type": "number" },
                  "total": { "type": "number" },
                  "customer": { "$ref": "#/components/schemas/Customer" },
                  "paymentMethod": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Order created with auto-generated invoice" } }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order details",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Order with transactions" } }
      },
      "patch": {
        "tags": ["Orders"],
        "summary": "Update order status (triggers notifications)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["pending", "paid", "shipped", "delivered", "cancelled", "refunded"] },
                  "tracking": { "type": "string" },
                  "carrier": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Order updated" } }
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Cancel an order",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Order cancelled" } }
      }
    },
...
  }
}